IRR: Optimize MessageWrapper
Problem: The message wrapper allocates a temporary buffer, reads
data into it and then copy the buffer to the response (string)
object. This is inefficient and twice as slow when sending big
messages such a DeltaPush/DeltaPreinstall which can be several
MiB big.

Solution: Resize the response object and write directely into it.

Bug: N/A
Test: None
Change-Id: I1cf5baacf346b544771250aed08b49b94ac5f13d