Fix ab-coverage targets query
This was broken by ag/5748911 which added an emulator dependency to a UI
test. Since coverage doesn't care about UI tests at the moment, this CL
just removes them from consideration in the query.

Additionally, it reads the return status of the query and will fail if
it fails. This will alert us if such a thing happens again.

Test: N/A
Bug: N/A
Change-Id: I6a114cd0209ba1988f16ce5f2180860b19b83f45