Update deploy pipeline to swap the dumped pids
Modifies ApkDumper to use the parsed APK information to select packages
to dump. A map of package --> process ids is passed to the ApkSwapper,
which determines how to perform the swap.

Modifies the installer swap command to accept a list of process ids,
instead of looking up the process ids on-device.

Scenarios validated:
- Code swap
- Code swap via debugger
- Code swap instrumented unit test
- Code swap instrumented unit test via debugger

Bug: 118495092
Test: ApkDumperTest + Manual
Change-Id: Ida54cf811b8b40ab83fdd304043a44fb02a12fbe