Fix CPU ART tracing after daemon refactoring.
CPU service used to hold a reference to 'agent_config',
which used to be a local variable in in perfd.cc:main().
Because main() is guaranteed to outlive CpuServiceImpl,
this is a correct and efficient implementation.

In Change Ib2740f287d834c6eda178a5376046ddfc293ee8e,
perfd.cc:main() was changed to Perfd::Initialize(),
which no longer outlives CpuServiceImpl. Therefore, the
reference to 'agent_config' becomes invalid after
Perfd::Initialize() finishes. Therefore, CPU service is
using an incorrect timeout value (zero (0))for ART
tracing, which breaks all ART tracing.

The fix is simple: change the reference to a copy.

Examined other use case of the local variable
'agent_config'. They are fine.

Bug: N/A (All CPU ART tracing is broken)
Test: N/A; try any CPU Java tracing.

Change-Id: I4d18d1ae20affa0df485d709b0e84b88428de856