Automatically rename Benchmarks with host-specific names
This reverts commit fc5b015c4b45043db99550a648cdfce6e143deff.

This will keep Benchmarks run on Mac and Windows from colliding with the
"default" Benchmarks, which are run on Linux. Changing the name of the
Benchmark object is sufficient to force data generated on Mac or Windows
hosts from appearing in the charts which already contain Linux
performance data.

In addition, this change updates Metric.java and MetricTest.java to
close some input/output streams that were previously being left open,
causing MetricTest to fail on Windows.

Bug: b/122838052
Test: Updated MetricTest and PerfDataTest
Test: Ran test locally and looked at perf_data.json
Change-Id: I7dc6646befcb6552cf3fa8c0f8618cb3a785326e