Introduce SdkComponents.
This CL introduce the new SdkComponents and hook it into GlobalScope.

SdkComponents is our new middleware that will be responsible for
shortcuting the SDK resolution to the directories where we would expect
to find it and then falling back to the traditional SdkHandler in case
we didn't manage to find it.

This will also be the main entry point for anything SDK related from our
tasks/gradle code, including being able to generate Gradle's Providers<>
for all the elements used in tasks, so we can move everything to be as
lazy as possible.

For now, this CL accomplishes:
* Moves all SDK access into SdkComponents (accesed via GlobalScope).
* Decouple totally the SDK getters from AndroidBuilder.
* Move some SDK access to Gradle's Providers, but not all of them.

Bug: 112700217
Test: existing
Change-Id: I501697faf40db61de386c172b26b339597cfb6b3