Allow user to request symlink to NDK
See comment in NdkSymlinker.kt for full description.

This change allows user to set a custom folder in local.properties
to symlink NDK to.

This replaces earlier method which attempts to symlink automatically.
The reason this way is better is that the user is aware of what's
happening and isn't surprised if he gets a permissions request
dialog on Windows.

Using this method I'm able to turn on five more integration tests on
Windows. This leaves only three remaining.

Bug: NA
Test: NdkSymlinker.kt is fully unittested. Also, I turned on five
      integration tests on Windows that rely on this feature.

Change-Id: I053128be0e7cd27516dc9f1a9b37eeedf638b8bb