Only add javac output as MergeJavaResourceTask input if there are APs
We only need to add javac output as input to MergeJavaResourceTask if
javac generates java resources, and that currently only happens if
there are annotation processors.

We check for annotation processors by checking for dependencies of the
annotationProcessor configuration, and add javac output to the
projectJavaRes input if we find such a dependency. This must be done via
a lazy file collection in case a dependency is added after the
configuration is resolved (e.g., via defaultDependencies).

Bug: NA
Test: MergeJavaResourceTaskTest
Change-Id: Icbf314a7a938fb1b2f6f09bac8fdec2dd90c0d19