Dex Artifact Transform: File dependency desugaring
When dexing with desugaring in artifact transforms, File dependencies
will not have any dependency metadata, and therefore cannot be correctly
desugared in an artifact transform.

This change moves the dexing with desugaring in to a separate task,
which allows them to be desugared against the whole classpath.

This should be replaced by a transform (b/128599004)

Test: Updated DesugarFileDependencyTest to reflect the fixed behaviour.
Bug: 128599004
Change-Id: I5f3b4a72eb2d282f5ca348826448208089fa592c