Update ddmlib to report push/install times
On a successful installation via the 'fallback' (non-delta push) path,
Apply Changes will now record the duration of the APK upload and the APK
installation, as well as the total duration of the install.

Note that for installations on API >= 21, the 'upload' time includes the
time to stream the APK to install-write, as well as the time taken by
install-write itself.

Fixes: 127660454
Test: DeployerRunnerTest
Change-Id: Idb30be7759ef86f4981c0e29588a0549561e490b
(cherry picked from commit 7be7150f530871822bf5407fd06e39b19d2b6817)