Disable runtime->compile classpath constraints for test/feature modules.
Modules that consume another APK module (so separate test modules and
dynamic feature modules) have a restricted view of the runtime classpath.
It contains the APK they consume but not the transitive dependencies
as we don't want to package the dependencies of the consumed APK
in the test app or in the dynamic features.

Because of this the runtime classpath of these modules is very limited
and therefore using it to set constraints on the compile classpath
does not work. We need to find a better way to synchronize these
classpaths.

For now, disable the constraint mechanism for these modules in
order to reduce issues in 3.5.

Bug: 132939682
Test: just existing test, proper fix in master will have actual scenario tests.
Change-Id: I4493041c89a53b8a5bde41db5f145ff5830698a3